Added support for changing the region end point, using the EC2 gem "server" key.
The endpoint defaults to us-east-1.ec2.amazonaws.com in "cloud_providers.aws.server_endpoint" in the rubber.yml config file.

Other than the change in the creation of the @ec2 object in Rubber::Cloud::Aws, there are no changes required anywhere else.

All of the tests pass, except for the final assertion in test/generator_test.rb (line 189) which fails, but this fails deep within the mocha gem, so I suspect it's to do with that, but I could be mistaken.

Please let me know if there's anything I've missed, or if I've headed off down the wrong path.