Access the compute provider through its accessor method and lazy init it.
Depending on the provider, fog may try to establish a connection at object creation time, rather than when making an actual request.  If the provider requires a VPN or some other special network configuration to make the connection, this can block normal deploys over SSH.  So, lazy initializing the object will guarantee we don't try to connect to the service until we need to make a request of that service.