Install iptables if it's not already installed and it's needed to set up firewall rules.
Most Ubuntu setups already have it installed, but I've run into issues with VMs used on VMware that don't. Rather than error out, we should just install it here. The problem is we can't make it a core_package because those packages are installed too late in the cycle. There's probably a better way to tidy this up in the future, but this works for now.