Be more graceful in failing to install core packages.
This may seem counterintuitive, since if a package is core and fails to install, that's usually bad. The problem is, since this list is hard-coded in, our users have no way of modifying it short of forking the gem. The real problem here is Ubuntu has a history of renaming packages that we need as core.  So, we have to include all possible names in our core list, but not all are guaranteed to install on the supported LTS releases.  We only need one of the alternatives to install properly, so in that scenario it is appropriate to ignore the failures.