Dupe the version string in the gemspec so a mutable copy is made.
We freeze the version string because it's supposed to be constant.  Older versions of RubyGems, such as the ones that ship with Ruby 1.9.3, expect to be able to modify the version string, however.  Generally the gemspec ships as part of the gem with the value resolved, so this isn't an issue.  But if bundler is using a git or local copy, where the version directly references our constant, its inability to modify the version results in a load error.

Fixes #453: Is VERSION.freeze needed?