Re-use established SSH connections.
When I introduced the mutex here to prevent threading issues, I pulled the connection out of the mutex so connections to multiple servers could be established at the same time.  This had the unintended consequence of establishing SSH connections repeatedly to the same server.  I'd still like to fix this to allow multiple servers to be connected to in parallel, but reducing single server connections from N to (N - 1) easily makes up for the lack of efficiency in the interim.  This may also fix problems with establishing a bunch of SSH connections that we never clean up properly.