Allow rubber:config execution to be optional again.
This fixes a regression introduced in bd82f2a9c15ddf2bd0d88821149a70b99ac03d7f by converting the rubber:config task to a nested namespace.

The reason this task needs to be optional is we support clusters consisting of machines that can't be deployed to.  In particular, we'll only deploy to Linux machines, currently.  Without this fix in place, any Windows or Mac machines in the cluster will also attempt to be deployed to and will end up failing.