Improved SSH key validation.
It's easy to end up with an empty key filename if using directory globbing.  Previously this wouldn't be caught because we'd do a File.expand_path on '', which just translates to the current working directory, which is a file that exists.  That tripped up our existing validation.  Now we check if the key_file value is empty in addition to being nil, preventing the problem altogether.